Dodanie indykatora temporary upgrade'u i zmienienie timePower zegarów na 1.5

+++ Assets/TimeHandler.cs
@@ -11,7 +11,7 @@ public class TimeHandler : MonoBehaviour
 
 
     #region Interface
-    public void addTime(int time)
+    public void addTime(float time)
     {
         totalTime += time;
     }
